requirements.txt
langchain
langchain-groq
pandas
streamlit
python-dotenv
openai
langchain-openai
langchain-core